g_drive/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
gunicorn==21.2.0
